[a1][mw/server][8/9] extract common logic into tcp util functions
Summary:

Test Plan:
